checkpoint: renames "Byte/EncodingCompatibilityError" to "Byte/Encoded/CompatibilityError"
