+++ src/library/Sequence/Byte/Encoded/CompatibilityError.ts
